pass_man/requirements.txt
# Django and Core Dependencies
Django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.7

# Authentication & Security
# JWT Authentication
djangorestframework-simplejwt==5.3.0
cryptography==41.0.7

# Environment Management
python-decouple==3.8

# Development Tools
django-extensions==3.2.3
django-debug-toolbar==4.2.0

# Testing
coverage==7.3.2
factory-boy==3.3.0

# Production Server
gunicorn==21.2.0

# Caching
redis==5.0.0
django-redis==5.3.0

# Static Files & Media
whitenoise==6.5.0

# Validation
django-phonenumber-field==7.1.0
phonenumbers==8.13.22

# Utilities
Pillow==10.0.1
python-dateutil==2.8.2
pytz==2023.3

# Security (Updated versions to resolve conflicts)
bandit==1.7.5
safety==3.0.1

# Code Quality (Updated versions to resolve conflicts)
black==24.3.0
flake8==7.0.0
isort==5.13.2
django-debug-toolbar==4.2.0
